zcash_pool_migration: gather the test fixtures into one module
`build::test_util` held the fixtures for this crate's unit tests, but
`engine`'s tests imported them too, so a helper set the engine depends on
was reached through `crate::build::test_util`. They move to
`testing::fixtures`, alongside the strategies, scenarios, and conformance
suite the crate already keeps there.

The new module is `cfg(test)` and is NOT re-exported from `testing`,
unlike its three siblings. It leans on `zcash_pool_migration_memory`,
which is a DEV-dependency, so it exists only for this crate's own tests
and must not reach the surface a downstream crate gets by enabling
`test-dependencies`.

Three helpers -- `account`, `spend_signability`, and
`assert_every_spend_is_identifiable` -- name no type this crate defines,
so they join the five siblings already re-exported from the memory crate
rather than being kept privately here.

`account_derivation` cannot follow them, and it is the boundary case that
shows where that line falls. It returns `AccountDerivation`, defined here,
and the memory crate links this crate as a library while these tests are a
second compilation of it, so the two types are distinct; re-exporting the
memory crate's copy fails to compile at `build_prep_tx` and
`build_transfer_pczt`. Every sibling that does move returns a type owned
by another crate, which is exactly why they can.

No public API is removed: the memory crate keeps `account_derivation` and
`without_account_derivation` even though neither has a caller outside it.
