+++ zcash_pool_migration/src/build.rs
@@ -186,86 +186,3 @@ pub(crate) fn output_action_index(
         .map(|i| i as u32)
         .ok_or_else(|| BuildError::Build(format!("no action index for output {output}")))
 }
-
-/// Test helpers shared by the split and transfer builders: a deterministic account, regtest network
-/// params, and a single-leaf Orchard witness. The seed-derived key, network, and witness helpers now
-/// live in the `zcash_pool_migration_memory` test-support crate and are re-exported here so both
-/// submodules' tests reuse them without change.
-#[cfg(test)]
-pub(crate) mod test_util {
-    use orchard::keys::FullViewingKey;
-
-    pub(crate) use zcash_pool_migration_memory::{
-        TARGET_HEIGHT, regtest_network, shared_anchor_witnesses, single_note_witness, spending_key,
-    };
-
-    /// An account's Orchard full viewing key (see [`spending_key`]).
-    pub(crate) fn account(seed: u64) -> FullViewingKey {
-        FullViewingKey::from(&spending_key(seed))
-    }
-
-    /// The ZIP 32 account derivation a wallet seeded with `seed` would report for the account
-    /// [`account`] views. Defined here rather than reused from `zcash_pool_migration_memory`
-    /// because that crate links this one as a library, so its `AccountDerivation` is a distinct
-    /// type from the one this crate's own test build sees.
-    pub(crate) fn account_derivation(seed: u64) -> super::AccountDerivation {
-        let mut seed_fingerprint = [0u8; 32];
-        seed_fingerprint[..8].copy_from_slice(&seed.to_le_bytes());
-        super::AccountDerivation::new(
-            zip32::fingerprint::SeedFingerprint::from_bytes(seed_fingerprint),
-            zip32::AccountId::ZERO,
-        )
-    }
-
-    /// How signable a built migration PCZT is, across both its Orchard and Ironwood bundles, as
-    /// `(needing a signature, of those, unidentifiable)`.
-    ///
-    /// An action needs a signature unless it is ALREADY SIGNED (a protocol padding dummy, which
-    /// the IO Finalizer signs with its throwaway key and then clears). One that needs a signature
-    /// is IDENTIFIABLE if it carries ZIP 32 derivation metadata, which is what lets an external
-    /// Signer that matches spends by derivation path recognize it as the account's. An action that
-    /// needs a signature and carries no derivation is one such a Signer skips as "not ours",
-    /// nothing else can sign, and whose transaction therefore fails to extract.
-    ///
-    /// The first count guards against a vacuous pass: a bundle of nothing but pre-signed dummies
-    /// has no unidentifiable spends while exercising nothing.
-    pub(crate) fn spend_signability(pczt: &pczt::Pczt) -> (usize, usize) {
-        fn count(bundle: &orchard::pczt::Bundle, unsigned: &mut usize, unidentifiable: &mut usize) {
-            for action in bundle.actions() {
-                if action.spend().spend_auth_sig().is_some() {
-                    continue;
-                }
-                *unsigned += 1;
-                if action.spend().zip32_derivation().is_none() {
-                    *unidentifiable += 1;
-                }
-            }
-        }
-
-        let (mut unsigned, mut unidentifiable) = (0, 0);
-        pczt::roles::verifier::Verifier::new(pczt.clone())
-            .with_orchard::<core::convert::Infallible, _>(|bundle| {
-                count(bundle, &mut unsigned, &mut unidentifiable);
-                Ok(())
-            })
-            .expect("the Orchard bundle parses")
-            .with_ironwood::<core::convert::Infallible, _>(|bundle| {
-                count(bundle, &mut unsigned, &mut unidentifiable);
-                Ok(())
-            })
-            .expect("the Ironwood bundle parses");
-        (unsigned, unidentifiable)
-    }
-
-    /// Asserts that every spend of `pczt` still needing a signature is identifiable to an external
-    /// Signer (see [`spend_signability`]), and returns how many such spends there are.
-    pub(crate) fn assert_every_spend_is_identifiable(pczt: &pczt::Pczt) -> usize {
-        let (unsigned, unidentifiable) = spend_signability(pczt);
-        assert_eq!(
-            unidentifiable, 0,
-            "{unidentifiable} of {unsigned} spends awaiting a signature carry no ZIP 32 \
-             derivation, so no external Signer can identify them as the account's",
-        );
-        unsigned
-    }
-}
